🚚 [Rename] 엔티티 클래스 및 관련 패키지 domain 하위로 이동
