+++ src/main/java/com/seohaeng/backend/domain/place/entity/Review.java
@@ -1,7 +1,6 @@
-package com.seohaeng.backend.review.entity;
+package com.seohaeng.backend.domain.place.entity;
 
-import com.seohaeng.backend.place.entity.Place;
-import com.seohaeng.backend.user.entity.User;
+import com.seohaeng.backend.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
