+++ src/main/java/com/seohaeng/backend/domain/readingSpot/entity/ReadingSpot.java
@@ -1,6 +1,6 @@
-package com.seohaeng.backend.readingSpot.entity;
+package com.seohaeng.backend.domain.readingSpot.entity;
 
-import com.seohaeng.backend.user.entity.User;
+import com.seohaeng.backend.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
