+++ src/main/java/com/seohaeng/backend/domain/readingSpot/entity/ReadingSpotComment.java
@@ -1,26 +1,27 @@
-package com.seohaeng.backend.notification.entity;
+package com.seohaeng.backend.domain.readingSpot.entity;
 
-import com.seohaeng.backend.user.entity.User;
+import com.seohaeng.backend.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Notification {
+public class ReadingSpotComment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String preview;
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "reading_spot_id", nullable = false)
+    private ReadingSpot readingSpot;
+
+    @Column(nullable = false)
+    private String content;
 
 }
