+++ src/main/java/com/seohaeng/backend/domain/travelCourse/entity/TravelCourse.java
@@ -1,6 +1,6 @@
 package com.seohaeng.backend.domain.travelCourse.entity;
 
-import com.seohaeng.backend.user.entity.User;
+import com.seohaeng.backend.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
